CRM_Backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-settings==2.8.1
python-jose[cryptography]==3.3.0
email-validator==2.2.0
python-multipart==0.0.20
pandas>=2.0
openpyxl>=3.1.0